My Solution to Grid-Layout-1
I learned that fr seems to be the usual measurement for grids. Also grids are much easier than flexboxes. I was able to use grid areas too, and it was intuitive. The only challenge I had here was finding the right measurement and how to implement them on the rows and columns.