Create website structure
I learned that we can use 0fr for giving space on my own. This only gives space depending on the content, meaning no extra padding on the row. Everything else was simple using grids. I learned how to use align-items for the footer, but I am currently having a hard time centering the text of each box and picture.